+++ client/src/reference-provider/parse.ts
@@ -31,7 +31,7 @@ function insertElementWithKey (templateName: string, fileLocation: VariablePathD
     }
 }
 
-function insertCalls (templateName: string, path: string, lineNrs: any[], allCallMaps: VariablePathMap) {
+function insertRefs (templateName: string, path: string, lineNrs: any[], allCallMaps: VariablePathMap) {
     lineNrs.forEach(lineItem => {
         const line = lineItem.line - 1;
 
@@ -48,30 +48,47 @@ function insertCalls (templateName: string, path: string, lineNrs: any[], allCal
     });
 }
 
-export function parseFile (file: string, allCallMaps: VariablePathMap) {
-    const documentText: string = fs.readFileSync(file, 'utf8');
-    const namespace: string = getIncludes(documentText);
-    const callPattern: RegExp = /\{(?:del)?call ([\w\d.]+)[^\w\d.].*/gm;
-    let m: RegExpExecArray;
+export function parseFile (file: string, allVariablePathMaps: VariablePathMap) {
+    const variablePattern: RegExp = /([\w\d._]+)[\w\d._, #]*/gim;
+    const content: string = fs.readFileSync(file, 'utf8');
+    let n: RegExpExecArray;
 
-    while (m = callPattern.exec(documentText)) {
-        const lineNr = linenumber(documentText, escapeRegExp(m[0]));
-        const template = m[0];
+    while (n = variablePattern.exec(content)) {
+        console.log(n)
+        const lineNr = linenumber(content, n[0]);
 
-        if (template.startsWith('.')) {
-            insertCalls(`${namespace}${template}`, file, lineNr, allCallMaps);
-        } else {
-            const alias: string = getMatchingVariable(template, documentText);
-
-            if (alias) {
-                const fullTemplatePath: string = normalizeAliasTemplate(alias, template);
-                insertCalls(fullTemplatePath, file, lineNr, allCallMaps);
-            } else {
-                insertCalls(template, file, lineNr, allCallMaps);
-            }
+        const alias: string = getMatchingVariable(n[0], content);
+        if (alias) {
+            insertRefs(n[1] || n[2], file, lineNr, allVariablePathMaps);
         }
+
+        insertRefs(n[1] || n[2], file, lineNr, allVariablePathMaps);
     }
 }
+// export function parseFile (file: string, allCallMaps: VariablePathMap) {
+//     const documentText: string = fs.readFileSync(file, 'utf8');
+//     const namespace: string = getIncludes(documentText);
+//     const callPattern: RegExp = /\{(?:del)?call ([\w\d.]+)[^\w\d.].*/gm;
+//     let m: RegExpExecArray;
+
+//     while (m = callPattern.exec(documentText)) {
+//         const lineNr = linenumber(documentText, escapeRegExp(m[0]));
+//         const template = m[0];
+
+//         if (template.startsWith('.')) {
+//             insertCalls(`${namespace}${template}`, file, lineNr, allCallMaps);
+//         } else {
+//             const alias: string = getMatchingVariable(template, documentText);
+
+//             if (alias) {
+//                 const fullTemplatePath: string = normalizeAliasTemplate(alias, template);
+//                 insertCalls(fullTemplatePath, file, lineNr, allCallMaps);
+//             } else {
+//                 insertCalls(template, file, lineNr, allCallMaps);
+//             }
+//         }
+//     }
+// }
 
 export function parseFilesForReferences (wsFolders: string[][]): VariablePathMap {
     const allCallMaps: VariablePathMap = {};